Adding grid system in prep for A* tutorial
GridLocation is a class that manages what is in a slot, whether you can pass or not.

Square grid is the class for the actual square grid. The math for different shaped grids is much more difficult so working with square grid to build up experience. Square grid has functions that get positions and draws the grid as well.

Added a grid to world.cs and initialized it.